perf(templates): compute contract publish impact in one query, not one per version
CheckContractPublishImpact dispatched a CheckTemplateVersionCompatibility
query (with its own SQL round-trip) for every published/archived version
on the old contract. referenced_paths now rides along on the existing
version query and compatibility is evaluated in-memory — every affected
row is on the old contract, so previousPublished.dataModel is each
version's old schema.

The comparison logic moves to a shared TemplateVersionCompatibilityEvaluator
used by both handlers; TemplateContractCompatibilityTest now exercises the
real evaluator instead of a hand-mirrored copy of the logic.

Closes #680
