+++ modules/epistola-core/src/test/kotlin/app/epistola/suite/templates/contracts/TemplateContractCompatibilityTest.kt
@@ -1,60 +1,27 @@
 package app.epistola.suite.templates.contracts
 
-import app.epistola.suite.templates.analysis.FieldIncompatibility
 import app.epistola.suite.templates.analysis.IncompatibilityReason
 import app.epistola.suite.templates.analysis.TemplateCompatibilityResult
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import tools.jackson.databind.ObjectMapper
 import tools.jackson.databind.node.ObjectNode
 
 /**
- * Tests the compatibility logic via SchemaPathNavigator directly.
- * The same logic is used by CheckTemplateVersionCompatibility query handler.
+ * Tests [TemplateVersionCompatibilityEvaluator] — the shared comparison used by both the
+ * CheckTemplateVersionCompatibility and CheckContractPublishImpact query handlers.
  */
 class TemplateContractCompatibilityTest {
 
-    private lateinit var navigator: SchemaPathNavigator
+    private val evaluator = TemplateVersionCompatibilityEvaluator()
     private val objectMapper = ObjectMapper()
-
-    @BeforeEach
-    fun setUp() {
-        navigator = SchemaPathNavigator()
-    }
 
     private fun schema(json: String): ObjectNode = objectMapper.readValue(json, ObjectNode::class.java)
 
     private val baseSchema = """{"type":"object","properties":{"name":{"type":"string"},"age":{"type":"integer"},"orders":{"type":"array","items":{"type":"object","properties":{"total":{"type":"number"},"status":{"type":"string"}}}}}}"""
 
-    /** Helper that mirrors the logic in CheckTemplateVersionCompatibilityHandler */
-    private fun checkCompatibility(paths: Set<String>, oldSchema: ObjectNode?, newSchema: ObjectNode?): TemplateCompatibilityResult {
-        if (paths.isEmpty()) return TemplateCompatibilityResult(compatible = true, incompatibilities = emptyList())
-        if (oldSchema != null && newSchema == null) {
-            return TemplateCompatibilityResult(
-                compatible = false,
-                incompatibilities = paths.map { FieldIncompatibility(it, IncompatibilityReason.FIELD_REMOVED, "Schema removed entirely") },
-            )
-        }
-        if (newSchema == null) return TemplateCompatibilityResult(compatible = true, incompatibilities = emptyList())
-
-        val incompatibilities = mutableListOf<FieldIncompatibility>()
-        for (path in paths) {
-            val newField = navigator.resolve(newSchema, path)
-            if (!newField.found) {
-                incompatibilities.add(FieldIncompatibility(path, IncompatibilityReason.FIELD_REMOVED, "\"$path\" not found in new schema"))
-                continue
-            }
-            if (oldSchema != null) {
-                val oldField = navigator.resolve(oldSchema, path)
-                if (oldField.found && oldField.type != newField.type) {
-                    incompatibilities.add(FieldIncompatibility(path, IncompatibilityReason.TYPE_CHANGED, "\"$path\" type changed from ${oldField.type} to ${newField.type}"))
-                }
-            }
-        }
-        return TemplateCompatibilityResult(compatible = incompatibilities.isEmpty(), incompatibilities = incompatibilities)
-    }
+    private fun checkCompatibility(paths: Set<String>, oldSchema: ObjectNode?, newSchema: ObjectNode?): TemplateCompatibilityResult = evaluator.evaluate(paths, oldSchema, newSchema)
 
     @Nested
     inner class Compatible {
